[UI] Log unsuccessful SPDA result as unsigned.
[UI] Log unsuccessful SetProcessDpiAwareness result as unsigned.

+++ src/xenia/ui/window_win.cc
@@ -66,7 +66,8 @@ bool Win32Window::OnCreate() {
       auto spda = (decltype(&SetProcessDpiAwareness))SetProcessDpiAwareness_;
       auto res = spda(PROCESS_PER_MONITOR_DPI_AWARE);
       if (res != S_OK) {
-        XELOGW("Failed to set process DPI awareness. (code = 0x{:08X})", res);
+        XELOGW("Failed to set process DPI awareness. (code = 0x{:08X})",
+               static_cast<uint32_t>(res));
       }
     }
